refactor(mc): extract FanChart, DepletionHistogram, PercentileTable
Pull the three reusable visual components out of MonteCarloTab into
standalone files keyed on MonteCarloResult fields. The composite stress
test (Task 6.2) will reuse these via MonteCarloResultView.

Snapshot baseline unchanged.
